Fix size/type ordering bug.
Hooray for testing against actual files!

Also experimented with several approaches to test buffer construction.
While we can't vec.append() yet, we can vec.extend() with other vecs,
which is much nicer than the 'for' loop. Still messy with strings
though.

One can use write! to copy a str into a vec, but it returns a result.
Failure options in this case aren't very interesting: allocation
failure is the only thing I can imagine, so we .unwrap() to avoid
the unused result warning.

It would be nicer to try!(write!()) but that requires a returned
Result to bubble errors, and the test harness doesn't permit this.
Since .unwrap() panics on error this is actually what we want for
the unit test.