Demonstrate recursion for content parsing in dump.rs.
On successful header parse, we match on the box type.
Although PartialEq is implemented between String and &str,
match doesn't seem to know how to convert if I just do
match a_string { "foo" => {}, ... } so instead we match
on a reference to a slice containing the whole string
contents.

If the box type is 'ftyp' we wrap the reader in a Take
so we can't read past the declared content size, and
pass that to the content reader function.

Otherwise, we skip the contents as before. There's a
problem with recursing here though: std::io::Seek
isn't implemented for Take<&mut File> so we can't
pass &mut content to anything which calls
skip_box_contents(). Not sure what to do there.