Move C api calls to a separate file.
This keeps things a little more tidy.

Apparently 'pub use capi::*' works in nightly, but not stable,
So we need to explicitly re-list exports for now.

Don't prefix 'use' paths in capi. The local crate doesn't have a name,
so we can can just use local paths here.

We can 'use' byteorder instead of 'extern crate'ing it
inside the module. It's already referenced from the top
of our crate.

Thanks to talchas and mbrubeck for help working this out.