Use namespaced 'use' to provide the Error trait.
std::io::Error implements the std::error::Error trait, but doesn't
re-export it. I don't understand why, but we need to import the
trait somehow to be able to call its description method, but we
can't 'use' both io::Error and error::Error since the names conflict.
As a work around we can provide an alternate namespace when we
import so the trait is available, but named something else. This
avoids the conflict and doesn't change the rest of our code since
the trait is only referenced indirectly.

Thanks to bluss in #rust for the suggestion.