Convert debug println!()s to log!() macros.
We have tracing println!()s throughout the code to help with
debugging and verification as we write, but this isn't very
nice for clients.

Add a 'trace' flag to the MediaContext struct, defaulting to
false.

Convert the tracing println!()s to a custom log!() macros
calls, which print the args only if context.trace is true.

This turns off the trace prints by default, but still lets
us use them for testing. This is the first rust macro I wrote,
so there may be some style issues