Pass Track to users to avoid having to fetch it from MediaContext.
This removes the awkward "if let Some(track) = ..." pattern all over the
place.

Also place an id on the Track for use with TrackScaledTime and
TrackTimeScale in place of the old index into MediaContext.tracks.
Ideally the track id would be a reference to the track, but I couldn't
figure out how to make that work with borrow checking.