Make CAPI error return codes positive.
These were only negative so we could overload the track count.

In gecko we have to flip the sign when recording error histograms
anyway, so it's easier if they're all positive.

It would be nice if we could align them with the tag values
used in-memory by the rust compiler, but that indexing starts
at zero, and we need that value for MP4PARSE_OK to match C
semantics.

N.B. this means one needs to check against MP4PARSE_OK instead of
just looking for a non-negative return code.