Add a test for parsing truncated (ftyp) boxes.
The existing design conflates expected UnexpectedEOFs with unexpected
UnexpectedEOFs.  We rely on an UnexpectedEOF at the end of a Take to
signal that we've finished reading the contents of a box and use this to
unwind the parser's stack.  In any other case, an UnexpectedEOF should
be treated as an error.

This is currently expected to fail.  I couldn't find a way to xfail a
test, so I've just set it to ignored for now.