capi: Use rustc --emit dep-info in examples/Makefile.
Instead of using a wildcard to pull in the module files,
use the rust compiler's dependency generation and include
it. Also remove some unnecessary code and duplication.

With rustc 1.6 we need to pass --crate-name to get
sensible dependency output. It doesn't propagate
the input filename from the link target.

Note that most makefiles avoid including the deps when
cleaning. I don't remember why this is necessary, but
it seems to work without it, so I've left that check
out for now.