Re-export stderr on build failure.
Invoking examples/Makefile from a custom build step was confusing.
Since it runs before the main build, any compilation failures in
the rust source would fail here, instead of in the later cargo
build step. However, since we capture stderr from the rustc
invocation to parse out the link dependencies, any error messages
are swallowed, leaving no information about what went wrong.

Piping to `tee` would be a good solution to this, so the output
is echoed as well as saved, but I couldn't find an equivalent to
the 2> stderr redirect which worked with pipes.

Instead, cat the saved rustc output back to stderr in the event
of a failure. This still displaces the syntax highlighting cargo
applies for normal build steps.