Fix a parser desync bug in read_stsd when reading multiple descriptions.
We forgot to wrap the box in a limit(), causing the iterator to start
consuming boxes from the parent box, which resulted in an UnexpectedEOF.

We used to treat multiple descriptions as a hard error (InvalidData), but
since they're present in bear-1280x720-v_frag-cenc_clear-all.mp4 and several
other Blink media test files, they need to be handled.  For now, take the
first description and just emit a log message and continue parsing for any
additional ones.