Detect non-BMFF audio SampleEntry and fail with a clearer error.
With some random .mp4 file on my machine (which turns out to have
major_brand="qt  " and compatible_brands=["qt  "], so it's really an old
QuickTime Movie file), the parser loses sync while trying to read the
audio SampleEntry.

This is because the AudioSampleEntry is a different format, with an
additional 16 bytes of data before the optional extra atoms, and with
the "esds" data wrapped inside a "wave" atom with a handful of others.

Existing Gecko builds play these files, so we likely need to handle them
correctly.

This probably reveals that we should be paying more attention to the
major and compatible brands we read out of the ftyp box.