Move env_logger to dev-dependencies.
When we switched to the `log` crate from our own custom logging
in v0.10, I added `env_logger` as an implementation so our code
examples would continue to work the same way.

However, since `env_logger` is _only_ needed for example code
is should be declared as a dev-dependency so it doesn't add
a transitive dependency to other builds using the library,
which may use a different logging implementation, or a just
a different version of `env_logger` than the one we've declared.