Parse tenc members related to pattern encryption.
Parse the tenc members for:
- crypt_byte_block: number of encrypted bytes in the pattern if pattern
encryption is used
- skip_byte_block: number of clear bytes in the pattern if pattern
encryption is used
- constant_iv: the iv that is used for a track by certain pattern based
encryption schemes, such as cbcs

Update the data structures used to contain new data. Note, the
is_encrypted member of the tenc structure has been changed from a u32 to
a u8 to reflect changes in the spec. This changes our interface.